Use a static national boundary file
This is because there is sometimes a delay between importing a new CCG and
importing its geographic boundary. During this period we are unable to
generate a complete national boundary from the CCG data and therefore
unable to infer practice boundaries. If we make this step part of the
import pipeline then the pipeline will be unable to complete
successfully during this period.

Given that the national boundary is not actually going to change when
CCG membership changes it is much more robust to store this in a file
that we can use independently from any CCG data.