Randomise order in which ports are used
Previously, attempting to re-run the functional tests locally after a
test run had recently completed would lead to a "port unavailable" error
because the port was still in a TIME_WAIT state. By shuffling the order
we make it very unlikely that subsequent runs will immediately re-use
the same port.