Avoid allocating new arrays when summing row groups
We already create a new array to hold our output values so we can sum
directly into this, rather than summing into a new array and then
copying the values across.

There's some annoying faff here around the differences between the
matrix and ndarray types, but we're going to have to deal with this
later in any case.