Test for mutliple concessions matching a single product
Because concessions are defined at the product-pack level it's possible
to have multiple concessions matching a single product. The existing
code doesn't handle this correctly and will double-count spending by
including the matching product twice.

In practice, this is rare and the effect is small. For instance, in our
current charts it affects only September and August 2018 and only by
around £50k within a total spend of about £9m.

Without prescribing at the product-pack level there's no perfect
solution here, but a reasonable approach is just to use whichever
concession gives the highest per-unit price.