Update the definition of "all practices"
Previously we included any practice which belonged to a "PCT" object,
regardless of whether that object was a CCG or some other thing. This
matched the original behaviour of the total_spending endpoint and of the
org_details endpoint when called without any organisation type.

However, this does not match the behaviour of the existing price
concessions calculation at the national level, where it only includes
"PCT" objects with an `org_type` of "CCG".

Given that national totals from the spending and org_details endpoints
were not actually used anywhere on the site, it seems acceptable to just
update the definition to match the price concessions calculation (which
is very publically displayed).