Refactor in-place add routine
Rather than always attempting an in-place add and falling back to
standard addition if anything goes wrong, we now only attempt it when we
expect it to work and we allow any resulting to errors to propogate.

If something changes, it seems better to find out early via an error
that our expectations are not being met then for it to result it a
silent slowdown of the MatrixSum operation.