Fix definition of "mean PPU" in price-per-unit calculation
The price-per-unit savings tool involves calculating the average
price-per-unit paid by an organisation for a given set of presentations
in a given month. Previously we were taking the average of ratios,
rather than the average ratio i.e. rather than doing `SUM(cost) /
SUM(quantity)` we were doing `AVG(cost / quantity)`.

This had the consequence that small prescriptions with an unusually high
or low price-per-unit would have a disproportionately large effect on
the overall PPU. As there is generally a floor on PPU but not a ceiling
(it's hard to pay less than a certain amount, but it's easy to pay very
high prices) this resulted in our mean PPU being higher than the true
value. This is turn meant that we were understating the savings
available through switching.

The difference turns out to be small but non-trivial. For instance, in
June 2019 the old method gave national savings of £22.9m and the
corrected method gives £24.3.