Add build_num_or_denom_fields
This will be used to set attributes on Measure related to numerators and
denominators, based on numerator_type/denominator_type.

This commit includes test_build_num_or_denom_fields (which will be
removed later) to check that the correct values of numerator_type/
denominator_type were set in a previous commit.