Fix CCG/practice bug in Analyse form bookmarks
The Analyse form allows the user to select all practices in a CCG by
supplying the CCG code in place of practice codes. This works correctly
the first time, however the function which repopulates the Analyse form
from URL parameters doesn't handle this correctly.

This PR fixes this bug and thereby closes #2058.