Fix incorrect aggregation SQL
In order to generate the correct SQL it's necessary that the definition
of `calc_value` come *after* the definition of `numerator` and
`denominator`. Previously, by luck of the order in which the keyword
arguments were processed, it worked to have all the definitions in one
`annotate` call. But the upgrade to Python 3 exposed the error here, so
now we use two `annotate` calls.