Move to using new URL syntax
This loses information about the regex that was previously used for
matching, but we never used that information.  The only effect this will
have is to change where 404s for invalid URLs get raised.  Rather than
being raised in the routing layer, they will now be raised by
get_object_or_404.

The one exception is the dmd_obj URL, where we need to keep requirement
that `id` is an integer to prevent it matching other URLs.