Fix CSV download URL on numerator breakdown page
Previously, the URLs were built seperately for the DataTable request and
for the download link. This made it easy to omit the `org_type`
parameter from the download link. In addition, there is some backwards
compatibility code which makes the org type optional and, if it is
omitted, tries to guess the type from the length of the code. So in
fact, the download link worked OK for practices and CCGs which is
probably why we never noticed the issue before. The addition of PCNs
made the problem more obvious.

The fix here is to generate the URL once, inside the view, and then use
it both places.