Include unsubscribe link in confirmation email
Just noticed while testing something that we don't do this (I think it
was basically impossible to do under the AllAuth regieme). Given that
we're not requiring confirmation any more I think it's more important
that we give people an easy way to unsubsribe in case they do so
accidentally or someone puts the wrong email address in.