Be more conservative about dropping indicies
I noticed this when trying to re-import all measure definitions (on
staging, luckily) to pick up the new tags that Brian had added and I
couldn't work out why it was taking so long.

This moves the option handling logic into the calling function, where I
feel it will be easier to spot cases like this when we make changes in
future.