Add "substitution sets" code
The role of this code is mostly explained in the docstring. This
centralises a bunch of logic previously spread between SQL queries,
Python code, and pre-computed BigQuery tables. This should also be the
only place in our PPU code which knows or cares about the structure of
BNF codes, so if we need to perform similar calculations using different
prescribing identifiers we hopefully have only one place to concentrate
on.