Add a basic caching decorator
This takes the total time to do the All England PPU calculation from
about 16s to about 3.5s and the time for a CCG calculation from ~14s to
~2s (all timings from my laptop).

From a cold cache it does increase the All England time to about 20s
(that's time spent writing to the cache) however this only needs to be
done once per new month of data and once the cache is populated any PPU
request for any organisation can use it. I think we can live with a
single slow request after a data import.

There's an additional layer of caching for the "total savings" function
as this figure is used on every dashboard page so we need it to be fast.
Without this second layer of caching, generating this figure takes about
3s; with caching it takes 0.0002s. This cache needs to be built once for
each organisation type but is shared between all organisations of the
same type.