Switch caching backend from Redis to DiskCache
The way the PPU code uses the cache (and the way the measures and NCSO
code will too, once we get round to it) involves storing lots of quite
large binary blobs of serialized matrices. Redis is pretty fast, but
shunting all this data to and fro over a socket feels a bit wasteful
compared to just grabbing a reference to it out of the kernel's page
cache as we do with the MatrixStore.

Fortunately someone has already done the hard work in creating a cache
backend that operates (or can be configured to operate) exactly like the
MatrixStore with the difference that it will start discarding old data
once the database reaches a certain size. That library is DiskCache:
http://www.grantjenks.com/docs/diskcache/index.html
https://github.com/grantjenks/python-diskcache

It looks pretty well battle and unit tested and profiling shows it to be
over twice as fast as Redis for our use case.

It's also one less external service to configure and manage (not that
Redis has given us much trouble). And to top it all off, it's *more*
SQLite, and who doesn't want that?

I have already updated the production and staging environment files to
add the lines:

  DISKCACHE_PATH=/mnt/database/diskcache/openprescribing
  DISKCACHE_PATH=/mnt/database/diskcache/openprescribing_staging