Refactor to aid cacheability
In order to make these functions cacheable we want to make them pure in
the sense that their values depend solely on the values of their
arguments and not on any global state. This means they need to be
refactored so all the global state they previously touched is now passed
in explicitly. This makes for slightly uglier and harder to follow code,
but I think the performance tradeoff is worth it.