Add standard_practice and standard_ccg definition
There is inconsistency in the codebase about where we restrict queries
just to setting 4 practices and where we don't. (I suspect the answer is
that we always want to restrict but sometimes we forgot and noone
noticed.) Unfortunately this inconsistency has worked its way into our
test fixtures and expected results so that fixing it would involve a lot
of fiddly work with the tests.

For now, therefore, I have created separate groupings to both include
and exclude non-standard practices. This allows us to exactly replicate
the existing behaviour and we can fix the inconsistencies as and when we
decide to rework the tests.