Add `cache_key` attribute to various objects
This is a small binary blob attached to the object which identifies its
state for caching purposes i.e. instances should have identical
cache_keys if and only if they have identical (relevant) states. This
will allow us to safely cache the values of functions which take these
kinds of objects as arguments.