Add function to get total savings for an organisation
We could in theory just loop over the results of `get_all_savings_for_orgs`
and add up the total but this involves doing work we don't need like
calcuation intermediate costs and PPUs. In addition, it makes sense to
calculate total savings for all orgs of a given type together so we can
cache the result.