Fix bubble API tests
This involves changing the expected names to match DM&D names and
changing quantities to be ints instead of floats.

The one substantive change is to the expected mean PPU in `test_focus`.
The previous code calculated the mean PPU from the "binned" (i.e rounded
to the nearest pence) PPU values which introduced a rounding error. The
new code calculates this correctly so that, in cases like this with a
single presentation, the "mean_ppu" and "plotline" values now match
exactly.