Update probiotics.json
This has been live as a pull request for awhile. I have amended "why it matters" to be clearer.

NB I have left notebook blank but there are a couple of [analysis here](https://github.com/ebmdatalab/jupyter-notebooks/tree/master/new_measures/Drafts/Draft%20VSL%233). I propose we tidy up this notebook while we change measure definition format and move measure notebooks to docker format in next fortnight. The measure is ok to deploy in advance of this and we can provide notebook if anyone asks in advance of tidy up.