Extract refresh_class_currency to own command
It now sets is_current to True on BNF classes if there has been any
prescribing in the past five years (rather than just the current month).
This is the behaviour we want, since is_current is used to determine
what can be searched for in the Analyse page.  This might be what the
code it replaces ended up doing, but we're being explicit about that
here.