Allow e2e tests to handle measures with no prescribing
Previous attempt to fix (7386a584) caused other problems... by setting
numerator_/denominator_where to `1 = 1`, all BNF codes are included in
the calculation for the numerator/denominator.  This means that the
numerator value can be much larger than the denominator value, which
doesn't make sense for measures where the numerator is a proportion of
the denominator.  When there's no prescribing, the numerator should be
zero.