Remove unused fields from prescribing schema
These fields are not present in the prescribing_v1 table in BQ.  It would
be possible, but slightly risky, to add them.  We may as well remove
them from the schema since they are not used.

The presence of these fields in the schema means that prescribing_v1 and
prescribing_v2 cannot be UNION ALLed together.