Exclude non-finite PPU values
When building the list of unique PPUs at which a drug has been obtained
we ignore the NaN values which result from practices which have not
prescribed that drug at all. However because of the issue where
quantities are occasionally rounded down to zero while cost remains
positive (#1373) we also occasionally get PPU values of `+inf` which
also need to be excluded. The `isfinite` check excludes both these and
NaN values.