Run cache garbage collection from a cron job
By default, DiskCache does this automatically as part of the request
which is convenient but has caused us database locking problems.

On both previous occasions I've seen the issue early and have been able
to fix it by stopping the server, deleting the cache, and restarting.
But we're unlikely to continue being so lucky so this feels worth fixing
now.