Mark some tests as expected failures
All the tests in `test_charts.py` fail in CI (BrowserStack running in Github
Actions) but pass locally (though a different set of tests fail locally and
pass in CI). Given that we were previously ignoring the functional test suite
altogether when running under Travis/Saucelabs and that we have no time or
funding to work on this, we've decided the least bad option is to mark these
tests as expected failures for now.

I've re-run the tests several times and they appear to at least be consistent
failures, unlike the intermittent ones we had with Travis/Saucelabs.