Remove Travis config file
Hopefully this will prevent Travis attempting to run on this repo.
There doesn't seem to be any per-repository setting to disable it, and I
don't want to deactivate it for our whole account as there are other
repos that use it.

This partially addresses #2828 but doesn't close it as I think there are
still other bits of Travis-related code that could do with cleaning up.