Use correct default for `include_in_alerts` measure config
At some point in past, all measures were included in alerts. We then
added an `include_in_alerts` config value which was set to `False` for
measures we didn't want included but omitted for all others. We then
realised that we weren't actually reading this config value anywhere
which we fixed in #2908 but got the default the wrong way round.

After this is merged we'll need to re-import the measure definitions
with:

    ./manage.py import_measures --definitions_only

Closes #2941