Handle missing drug tariff data for a month
In this case, even if we have price concession data for this month we
can't use that data because we don't have the standard tariff price
against which to compare it.

I don't know if we _should_ have tariff data and something has broken
with the import. But either way this feels like something we should
handle gracefully.

Fixes #3059