Remove another broken YouTube embed
This is the same underlyng issue identified in #3190 and fixed in #3193
whereby a previously public video is now private and no one has
credentials for the account which owns it. However this is a different
video from the one previously on the homepage.

I've grepped for YouTube embeds and confirmed that the remaining videos
are playable. Presumably because they all belong to the EBMDataLab
account:
https://www.youtube.com/channel/UC4V1SB4aV2-7zvzhyI98b4Q/videos