Remove Ansible scripts
These install Python 3.7 and we run 3.8 in production. We've just
noticed this because some of our dependencies are dropping 3.7 support
and thus CI is failing.

It turns out to be non-trivial to get the scripts to install 3.7 because
there's no package for it in the OS version we're using. At present no
one who works on OpenPrescribing is using the Ansible/Vagrant set up so
it's not worth trying to fix this right now.