Add project and dataset placeholders to measure SQL
We check for the project placeholder ("{project}" ) to ensure that we
don't hardcode "embdatalab" as the project name, and so the project
placeholder must be present in any SQL that is run against BQ.  (This is
not failsafe, but it's a good reminder.)

When running against the "ebmdatalabtest" project, we give each test run
its own set of datasets, identified by a nonce.  The full names of the
dataset (eg "hscic_1234") replaces the dataset placeholders.

Fixes #3454.