Make workaround for implicit int conversion more robust
If you attempt to use JSON keys which are strings that can be parsed as
ints, Django will automatically convert them to ints:
https://github.com/django/django/blob/0b8a0296/django/db/models/fields/json.py#L311-L314

This makes it impossible to work with the JSON structures we have in the
database which have keys like `"10"` (as distinct from `10`) which we
have no means of addressing.

We already knew about this and have a workaround in the form of the
`NotAnInt()` class which is just a `str` with its `__int__` method
overridden to throw an error.

However somewhere in the changes between Django 3.0 and 3.1, this value
is now getting passed through `str` _before_ it is checked for being an
int. This converted a `NotAnInt` into a plain `str` and undid the
workaround.

We workaround the undoing of the workaround by overloading `__str__` to
return the original `NotAnInt`.

And this, kids, is why we can't have nice things.