Raise explicit error if no data available for date
Previously this would end up resulting in a `KeyError`. You'd have to
manually edit the date in the URL to trigger this, but someone had
obviously done this and was trying repeatedly and the Sentry alerts were
getting annoying.

This doesn't handle requesting dates too far in the past, but let's
worry about that when it becomes a problem.