Tweak the format of NCSO importer Slack report
Having seen this in the wild I now realise that (a) it's not fixed-width
formatted and (b) the kind of report that's helpful when doing on big
re-import of everything using a new system is not the same as what's
helpful when doing frequent incremental imports. In particular,
constantly being told about mismatched VMPPs from historical
publications is just noise.