Stop filtering practices with empty list sizes
When calculating monthly/annual totals, the Analyse page includes
prescribing from practices which don't have any recorded list size.
However it doesn't display such practices on the chart, which leads to
confusion.

The reason it doesn't display such practices is that the MatrixStore
powered API goes to some lengths to mimic exactly what the Postgres
powered API did. However I suspect this behaviour was just a consequence
of the way the SQL happened to work rather than a deliberate decision.
And, given that it's confusing, we should stop doing it.