Drop assertions about how many measure objects were created by e2e tests
For a handful of measures (eg if both numerator and denominator join on
empty dm+d tables) there is not enough test data to create any measure
objects.  Rather than special casing certain measures (eg ghost branded
generics) as we have done in the past, we don't make any assertions
about how many objects we expect to be created.

It's still valuable to know that the pipeline runs to completion without
crashing.