Change GBG download link
This changes the download link to download link to retrive exactly the
data shown in the table, rather than additionally breaking the down by
practice. Breaking down by practice for large CCGs/SICBLs was too slow
to be practical and led to timeouts.

As a consequence, we change the button text to "Download data" rather
than "Download raw data".

Closes #4600