Handle absence of Google Analytics `ga()` function
Well, not so much handle its absence as ensure its presence as that is
by far the easier "fix" here.

Without this, the chart images needed for "search bookmark" (aka
Analyse) alerts don't render.

I'm not exactly sure when or why this stopped working. Possibly it was
as a consequence of #4343, or possibly it was already broken and that
point and that PR just failed to fix it.

I've confirmed this works locally, but it's hard to be 100% sure until
we try it in prod.