Add a `--quiet` flag to NCSO alerts command
For reasons that I don't have time to investigate it seems that both
stdout _and_ stderr get reported back to Slack by BennettBot. This means
that all the helpful logging I added ends up Slack, making the command
extremely noisy.

This PR adds a `--quiet` flag (which I will separately add to the
BennettBot invocation) to turn off the noisy logging. I want to keep the
logging on by default though so that it appears when you run the command
manually.