Restore original BNF map behaviour
This undoes the attempt to let later files remove mappings supplied in
earlier files. This totally failed to work because the files are
iterated in reverse order of date. I've tried to work out _why_ they are
applied in reverse order (intuitively, this doesn't make sense to me)
but I haven't been able to understand this.

We retain the separation of parse and modify steps, and keep the
modifications in a single transaction as these seem like sensible
robustness improvements which shouldn't change the happy path behaviour
of the code.