Remove staging service files
We've long since removed the staging web service, but because the
service files stuck around systemd was constantly trying to start the
service, reaslising it didn't have the appropriate configuration and
then exiting. Every four seconds. If nothing else, this clutters up the
logs which I noticed when trying to work out why a cronjob wasn't
running.