Fix logic for selecting practices by SICBL
The API allows you to retrieve practice-level data by supplying a
3-character code for a SICBL/CGG which then selects all the practices
belonging to it. However, previously, if you supplied the code of a
closed SICBL with no practices it would end up returning five years'
worth of data for every practice in England. Not only is this an
incorrect respond but it takes a lot of server resources to generate and
has the side effect of segfaulting PhantomJS when it tries to render the
chart image.

This PR applies the smallest fix to prevent it doing this.