Disable uploading supplementary tables on preview
This didn't really work as intended because the version of the tables
uploaded came from the `main` branch rather than the branch being
previewed. But we can't just upload the preview version because it would
overwrite the real version (where that already exists). And we can't
just upload it under a new name because then the measure SQL would ned
rewriting to refer to the new table.

The current workflow is to upload data manually in BQ while previewing
and then commit to git once satisfied it's correct.

Slack thread:
https://bennettoxford.slack.com/archives/C051A4P27GE/p1737365651665029